proctests: Move proc service related things into separate header
